Some refactoring and changed all int to int64

+++ btcturk/order_book.go
@@ -2,14 +2,14 @@ package btcturk
 
 import "fmt"
 
-// https://docs.btcturk.com/#order-book
+// OrderBook https://docs.btcturk.com/#order-book
 type OrderBook struct {
-	TimeStamp float64    `json:"timestamp"`
+	TimeStamp int64      `json:"timestamp"`
 	Bids      [][]string `json:"bids"`
 	Asks      [][]string `json:"asks"`
 }
 
-// GET ?pairSymbol=BTC_TRY
+// OrderBook GET ?pairSymbol=BTC_TRY
 // Or
 // GET ?pairSymbol=BTC_TRY&limit=100
 func (c *Client) OrderBook() (OrderBook, error) {
